chore: move BaseTransformer for its own file
