+++ nlpiper/transformers/cleaners.py
@@ -6,7 +6,7 @@ from typing import Optional
 from unicodedata import normalize, combining
 
 from nlpiper.core import Document
-from nlpiper.transformers import (
+from nlpiper.transformers.base import (
     BaseTransformer,
     TransformersType,
     add_step,
